# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ledger of Decisions API",
        "description": "決策驅動的記帳 API - 每筆消費都標註決策意圖",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "本地開發環境"
        }
    ],
    "paths": {
...
        "/statistics/intents": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "Intent 統計",
                "description": "查看各決策意圖的出現次數與平均信心程度",
                "operationId": "e7e5dbea28c297168539449bc689019b",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "起始日期",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "結束日期",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "preset",
                        "in": "query",
                        "description": "預設時間範圍",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "today",
                                "this_week",
                                "this_month"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/statistics/summary": {
            "get": {
                "tags": [
                    "Statistics"
                ],
                "summary": "消費總覽",
                "description": "取得消費統計總覽，包含分類統計、意圖統計、衝動消費比例",
                "operationId": "5ce0d2094febdb98df917efd9114b3a9",
                "parameters": [
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "起始日期",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "結束日期",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "preset",
                        "in": "query",
                        "description": "預設時間範圍",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "today",
                                "this_week",
                                "this_month"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
    },
...
}